requirements.txt
black==23.12.1
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.24
tqdm==4.66.1